Code Will mardi matin
- Implémentation de InterventionDAO
- Changement de tous les types Date
- Implémentation de certains services dans servicePersonne (afficherOpeDuJour, afficheHistorique, avertirEmploye)
- Notation des TODO dans le code pour voir ce qu'il reste à faire

+++ TPDASI/TP_DASI/src/main/java/dao/PersonneDAO.java
@@ -35,7 +35,7 @@ public class PersonneDAO {
         obtenirEntityManager().merge(cli);
     }
     
-    public static Personne findPersonneByIndex (int index) {
+    public static Personne findPersonneByIndex (Long index) {
         return obtenirEntityManager().find(Employe.class, index);
     }
